Struct space_net::boot_node::config::Config

pub struct Config {

    pub(crate) id: ZenohId,
    pub(crate) mode: Option<WhatAmI>,
    pub connect: ConnectConfig,
    pub listen: ListenConfig,
    pub scouting: ScoutingConf,
    pub timestamping: TimestampingConf,
    pub(crate) queries_default_timeout: Option<u64>,
    pub routing: RoutingConf,
    pub aggregation: AggregationConf,
    pub transport: TransportConf,
    pub adminspace: AdminSpaceConf,
    pub(crate) plugins_search_dirs: Vec<String, Global>,
    pub(crate) plugins: PluginsConfig,

}

The main configuration structure for Zenoh.

Most fields are optional as a way to keep defaults flexible. Some of the fields have different default values depending on the rest of the configuration.

To construct a configuration, we advise that you use a configuration file (JSON, JSON5 and YAML are currently supported, please use the proper extension for your format as the deserializer will be picked according to it).

Fields
id: ZenohIdmode: Option<WhatAmI>connect: ConnectConfig

Which zenoh nodes to connect to.

listen: ListenConfig

Which endpoints to listen on. zenohd will add tcp/[::]:7447 to these locators if left empty.

scouting: ScoutingConftimestamping: TimestampingConf

Configuration of data messages timestamps management.

queries_default_timeout: Option<u64>routing: RoutingConf

The routing strategy to use and it’s configuration.

aggregation: AggregationConf

The declarations aggregation strategy.

transport: TransportConfadminspace: AdminSpaceConf

Configuration of the admin space.

plugins_search_dirs: Vec<String, Global>plugins: PluginsConfig
Implementations

impl Config

pub fn validate(&self) -> bool

pub fn new(
id: ZenohId,
mode: Option<WhatAmI>,
connect: ConnectConfig,
listen: ListenConfig,
scouting: ScoutingConf,
timestamping: TimestampingConf,
queries_default_timeout: Option<u64>,
routing: RoutingConf,
aggregation: AggregationConf,
transport: TransportConf,
adminspace: AdminSpaceConf,
plugins_search_dirs: Vec<String, Global>,
plugins: PluginsConfig
) -> Result<Config, Config>

pub fn id(&self) -> &ZenohId

The Zenoh ID of the instance. This ID MUST be unique throughout your Zenoh infrastructure and cannot exceed 16 bytes of length. If left unset, a random UUIDv4 will be generated.

pub fn set_id(&mut self, value: ZenohId) -> Result<ZenohId, ZenohId>

The Zenoh ID of the instance. This ID MUST be unique throughout your Zenoh infrastructure and cannot exceed 16 bytes of length. If left unset, a random UUIDv4 will be generated.

pub fn mode(&self) -> &Option<WhatAmI>

The node’s mode (“router” (default value in zenohd), “peer” or “client”).

pub fn set_mode(
&mut self,
value: Option<WhatAmI>
) -> Result<Option<WhatAmI>, Option<WhatAmI>>

The node’s mode (“router” (default value in zenohd), “peer” or “client”).

pub fn connect(&self) -> &ConnectConfig

Which zenoh nodes to connect to.

pub fn set_connect(
&mut self,
value: ConnectConfig
) -> Result<ConnectConfig, ConnectConfig>

Which zenoh nodes to connect to.

pub fn listen(&self) -> &ListenConfig

Which endpoints to listen on. zenohd will add tcp/[::]:7447 to these locators if left empty.

pub fn set_listen(
&mut self,
value: ListenConfig
) -> Result<ListenConfig, ListenConfig>

Which endpoints to listen on. zenohd will add tcp/[::]:7447 to these locators if left empty.

pub fn scouting(&self) -> &ScoutingConf

pub fn set_scouting(
&mut self,
value: ScoutingConf
) -> Result<ScoutingConf, ScoutingConf>

pub fn timestamping(&self) -> &TimestampingConf

Configuration of data messages timestamps management.

pub fn set_timestamping(
&mut self,
value: TimestampingConf
) -> Result<TimestampingConf, TimestampingConf>

Configuration of data messages timestamps management.

pub fn queries_default_timeout(&self) -> &Option<u64>

The default timeout to apply to queries in milliseconds.

pub fn set_queries_default_timeout(
&mut self,
value: Option<u64>
) -> Result<Option<u64>, Option<u64>>

The default timeout to apply to queries in milliseconds.

pub fn routing(&self) -> &RoutingConf

The routing strategy to use and it’s configuration.

pub fn set_routing(
&mut self,
value: RoutingConf
) -> Result<RoutingConf, RoutingConf>

The routing strategy to use and it’s configuration.

pub fn aggregation(&self) -> &AggregationConf

The declarations aggregation strategy.

pub fn set_aggregation(
&mut self,
value: AggregationConf
) -> Result<AggregationConf, AggregationConf>

The declarations aggregation strategy.

pub fn transport(&self) -> &TransportConf

pub fn set_transport(
&mut self,
value: TransportConf
) -> Result<TransportConf, TransportConf>

pub fn adminspace(&self) -> &AdminSpaceConf

Configuration of the admin space.

pub fn set_adminspace(
&mut self,
value: AdminSpaceConf
) -> Result<AdminSpaceConf, AdminSpaceConf>

Configuration of the admin space.

pub fn plugins_search_dirs(&self) -> &Vec<String, Global>

A list of directories where plugins may be searched for if no __path__ was specified for them. The executable’s current directory will be added to the search paths.

pub fn set_plugins_search_dirs(
&mut self,
value: Vec<String, Global>
) -> Result<Vec<String, Global>, Vec<String, Global>>

A list of directories where plugins may be searched for if no __path__ was specified for them. The executable’s current directory will be added to the search paths.

pub fn plugins(&self) -> &PluginsConfig

The configuration for plugins.

Please refer to PluginsConfig’s documentation for further details.

pub fn set_plugins(
&mut self,
value: PluginsConfig
) -> Result<PluginsConfig, PluginsConfig>

The configuration for plugins.

Please refer to PluginsConfig’s documentation for further details.

impl Config

pub fn from_deserializer<'d, D>(
d: D
) -> Result<Config, Result<Config, <D as Deserializer<'d>>::Error>>where
D: Deserializer<'d>,
Config: Deserialize<'d>,

impl Config

pub fn add_plugin_validator(
&mut self,
name: impl Into<String>,
validator: Arc<dyn Fn(&str, &Map<String, Value>, &Map<String, Value>) -> Result<Option<Map<String, Value>>, Box<dyn Error + Sync + Send + 'static, Global>> + Sync + Send + 'static>
)

pub fn plugin(&self, name: &str) -> Option<&Value>

pub fn sift_privates(&self) -> Config

pub fn remove<K>(
&mut self,
key: K
) -> Result<(), Box<dyn Error + Sync + Send + 'static, Global>>where
K: AsRef<str>,

impl Config

pub fn from_file<P>(
path: P
) -> Result<Config, Box<dyn Error + Sync + Send + 'static, Global>>where
P: AsRef<Path>,

pub fn libloader(&self) -> LibLoader