Trait space_net::boot_node::SplitBuffer

pub trait SplitBuffer<'a> {
    type Slices: Iterator<Item = &'a [u8]> + ExactSizeIterator;

    fn slices(&'a self) -> Self::Slices;

    fn is_empty(&'a self) -> bool { ... }
    fn len(&'a self) -> usize { ... }
    fn contiguous(&'a self) -> Cow<'a, [u8]> { ... }
}

A trait for buffers that can be composed of multiple non contiguous slices.

Required Associated Types

type Slices: Iterator<Item = &'a [u8]> + ExactSizeIterator

Required Methods

fn slices(&'a self) -> Self::Slices

Gets all the slices of this buffer.

Provided Methods

fn is_empty(&'a self) -> bool

Returns true if the buffer has a length of 0.

fn len(&'a self) -> usize

Returns the number of bytes in the buffer.

fn contiguous(&'a self) -> Cow<'a, [u8]>

Returns all the bytes of this buffer in a conitguous slice. This may require allocation and copy if the original buffer is not contiguous.