Struct space_net::node::config::ScoutingConf

pub struct ScoutingConf {
    pub(crate) timeout: Option<u64>,
    pub(crate) delay: Option<u64>,
    pub multicast: ScoutingMulticastConf,
    pub gossip: GossipConf,
}

Fields
timeout: Option<u64>delay: Option<u64>multicast: ScoutingMulticastConf

The multicast scouting configuration.

gossip: GossipConf

The gossip scouting configuration.

Implementations

impl ScoutingConf

pub fn validate(&self) -> bool

pub fn new( timeout: Option<u64>, delay: Option<u64>, multicast: ScoutingMulticastConf, gossip: GossipConf ) -> Result<ScoutingConf, ScoutingConf>

pub fn timeout(&self) -> &Option<u64>

In client mode, the period dedicated to scouting for a router before failing. In milliseconds.

pub fn set_timeout( &mut self, value: Option<u64> ) -> Result<Option<u64>, Option<u64>>

In client mode, the period dedicated to scouting for a router before failing. In milliseconds.

pub fn delay(&self) -> &Option<u64>

In peer mode, the period dedicated to scouting remote peers before attempting other operations. In milliseconds.

pub fn set_delay( &mut self, value: Option<u64> ) -> Result<Option<u64>, Option<u64>>

In peer mode, the period dedicated to scouting remote peers before attempting other operations. In milliseconds.

pub fn multicast(&self) -> &ScoutingMulticastConf

The multicast scouting configuration.

pub fn set_multicast( &mut self, value: ScoutingMulticastConf ) -> Result<ScoutingMulticastConf, ScoutingMulticastConf>

The multicast scouting configuration.

pub fn gossip(&self) -> &GossipConf

The gossip scouting configuration.

pub fn set_gossip(&mut self, value: GossipConf) -> Result<GossipConf, GossipConf>

The gossip scouting configuration.

impl ScoutingConf

pub fn from_deserializer<'d, D>( d: D ) -> Result<ScoutingConf, Result<ScoutingConf, <D as Deserializer<'d>>::Error>>where D: Deserializer<'d>, ScoutingConf: Deserialize<'d>,