Struct space_net::node::Node

pub struct Node {
    pub(crate) session: Arc<Session>,
    pub(crate) cluster_name: String,
    pub(crate) zid: String,
    pub(crate) api_requester_tx: Sender<ApiMessage>,
    pub(crate) api_responder_rx: Receiver<ApiResponse>,
}

Fields
session: Arc<Session>cluster_name: Stringzid: Stringapi_requester_tx: Sender<ApiMessage>api_responder_rx: Receiver<ApiResponse>
Implementations

impl Node

pub fn new(cluster_name: &str) -> Self

pub fn join(&mut self, site: (f64, f64))

pub fn leave(&mut self)

Node is dropped and leaves the cluster when not busy with task.

pub fn leave_on_pressed(&self, key: char)

End node when the user presses a key. Node is dropped and leaves the cluster. use to return Self but cant use builder pattern in C

pub fn get_zid(&self) -> String

Get the zid of the node

pub fn get_status(&self) -> NodeStatus

Get node status

fn set_status(&self, status: NodeStatus)

fn set_site(&self, site: (f64, f64))

pub fn get_site(&self) -> (f64, f64)

pub fn closest_neighbour(&self, point: (f64, f64)) -> String

pub fn send_message(&self, payload: Vec<u8>, topic: &str)

pub fn get_neighbours_sites(&self) -> Vec<(String, (f64, f64))>

Get the neighbours of the node

pub fn get_neighbours(&self) -> Vec<String>

pub fn is_neighbour(&self, zid: &str) -> bool

Check if the node is a neighbour

pub fn get_polygon(&self) -> Vec<(f64, f64)>

Get the polygon of the node

pub fn is_in_polygon(&self, point: (f64, f64)) -> bool

Check if the point site is in the polygon. Ray casting algorithm.