Struct space_net::node::Selector

#[non_exhaustive]
pub struct Selector<'a> {
    pub key_expr: KeyExpr<'a>,
    pub(crate) parameters: Cow<'a, str>,
}

A selector is the combination of a Key Expression, which defines the set of keys that are relevant to an operation, and a set of parameters with a few intendend uses:

specifying arguments to a queryable, allowing the passing of Remote Procedure Call parameters

filtering by value,

filtering by metadata, such as the timestamp of a value,

specifying arguments to zenoh when using the REST API.

When in string form, selectors look a lot like a URI, with similar semantics:

the key_expr before the first ? must be a valid key expression.

the parameters after the first ? should be encoded like the query section of a URL:

parameters are separated by &,

the parameter name and value are separated by the first =,

in the absence of =, the parameter value is considered to be the empty string,

both name and value should use percent-encoding to escape characters,

defining a value for the same parameter name twice is considered undefined behavior, with the encouraged behaviour being to reject operations when a duplicate parameter is detected.

Zenoh intends to standardize the usage of a set of parameter names. To avoid conflicting with RPC parameters, the Zenoh team has settled on reserving the set of parameter names that start with non-alphanumeric characters.

The full specification for selectors is available here, it includes standardized parameters.

Queryable implementers are encouraged to prefer these standardized parameter names when implementing their associated features, and to prefix their own parameter names to avoid having conflicting parameter names with other queryables.

Here are the currently standardized parameters for Zenoh (check the specification page for the exhaustive list):

_time: used to express interest in only values dated within a certain time range, values for this parameter must be readable by the Zenoh Time DSL for the value to be considered valid.

[unstable] _anyke: used in queries to express interest in replies coming from any key expression. By default, only replies whose key expression match query’s key expression are accepted. _anyke disables the query-reply key expression matching check.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

key_expr: KeyExpr<'a>

The part of this selector identifying which keys should be part of the selection.

parameters: Cow<'a, str>
Implementations

impl<'a> Selector<'a>

pub fn parameters(&self) -> &str

Gets the parameters as a raw string.

pub fn parameters_map<K, V>( &'a self ) -> Result<HashMap<K, V, RandomState>, Box<dyn Error + Send + Sync + 'static, Global>>where K: AsRef<str> + Hash + Eq, <<<Selector<'a> as Parameters<'a>>::Decoder as Iterator>::Item as Parameter>::Name: Into<K>, <<<Selector<'a> as Parameters<'a>>::Decoder as Iterator>::Item as Parameter>::Value: Into<V>,

Extracts the selector parameters into a hashmap, returning an error in case of duplicated parameter names.

pub fn parameters_cowmap( &'a self ) -> Result<HashMap<Cow<'a, str>, Cow<'a, str>, RandomState>, Box<dyn Error + Send + Sync + 'static, Global>>

Extracts the selector parameters’ name-value pairs into a hashmap, returning an error in case of duplicated parameters.

pub fn parameters_stringmap( &'a self ) -> Result<HashMap<String, String, RandomState>, Box<dyn Error + Send + Sync + 'static, Global>>

Extracts the selector parameters’ name-value pairs into a hashmap, returning an error in case of duplicated parameters.

pub fn parameters_mut(&mut self) -> &mut String

Gets a mutable reference to the parameters as a String.

Note that calling this function may cause an allocation and copy if the selector’s parameters wasn’t already owned by self. self owns its parameters as soon as this function returns.

pub fn set_parameters(&mut self, selector: impl Into<Cow<'a, str>>)

pub fn borrowing_clone(&'a self) -> Selector<'a>

pub fn into_owned(self) -> Selector<'static>

pub fn to_owned(&self) -> Selector<'static>

👎Deprecated: If you have ownership of this selector, prefer Selector::into_owned

pub fn split(self) -> (KeyExpr<'a>, Cow<'a, str>)

Returns this selectors components as a tuple.

pub fn with_parameters(self, parameters: &'a str) -> Selector<'a>

Sets the parameters part of this Selector.

pub fn extend<'b, I, K, V>(&'b mut self, parameters: I)where I: IntoIterator, <I as IntoIterator>::Item: Borrow<(K, V)>, K: AsRef<str> + 'b, V: AsRef<str> + 'b,

pub fn with_time_range(&mut self, time_range: TimeRange<TimeExpr>)

Sets the time range targeted by the selector.

pub fn remove_time_range(&mut self)