Trait space_net::node::IntoCallbackReceiverPair

pub trait IntoCallbackReceiverPair<'a, T> {
    type Receiver;

    // Required method
    fn into_cb_receiver_pair(
        self
    ) -> (Arc<dyn Fn(T) + Send + Sync + 'a>, Self::Receiver);
}

A type that can be converted into a [Callback]-receiver pair.

When Zenoh functions accept types that implement these, it intends to use the [Callback] as just that, while granting you access to the receiver through the returned value via std::ops::Deref and std::ops::DerefMut.

Any closure that accepts T can be converted into a pair of itself and ().

Required Associated Types

type Receiver

Required Methods

fn into_cb_receiver_pair( self ) -> (Arc<dyn Fn(T) + Send + Sync + 'a>, Self::Receiver)