Trait space_net::node::Parameters

pub trait Parameters<'a> {
    type Decoder: Iterator + 'a;

    // Required method
    fn decode(&'a self) -> Self::Decoder
       where <Self::Decoder as Iterator>::Item: Parameter;

    // Provided methods
    fn decode_into_map<N, V>(
        &'a self
    ) -> Result<HashMap<N, V, RandomState>, Box<dyn Error + Send + Sync + 'static, Global>>
       where <Self::Decoder as Iterator>::Item: Parameter,
             N: AsRef<str> + Hash + Eq,
             <<Self::Decoder as Iterator>::Item as Parameter>::Name: Into<N>,
             <<Self::Decoder as Iterator>::Item as Parameter>::Value: Into<V> { ... }
    fn get_parameters<const N: usize>(
        &'a self,
        names: [&str; N]
    ) -> Result<[Option<<<Self::Decoder as Iterator>::Item as Parameter>::Value>; N], Box<dyn Error + Send + Sync + 'static, Global>>
       where <Self::Decoder as Iterator>::Item: Parameter { ... }
    fn get_bools<const N: usize>(
        &'a self,
        names: [&str; N]
    ) -> Result<[bool; N], Box<dyn Error + Send + Sync + 'static, Global>>
       where <Self::Decoder as Iterator>::Item: Parameter { ... }
    fn time_range(
        &'a self
    ) -> Result<Option<TimeRange<TimeExpr>>, Box<dyn Error + Send + Sync + 'static, Global>>
       where <Self::Decoder as Iterator>::Item: Parameter { ... }
}

A trait to help decode zenoh selector parameters.

Most methods will return an Error if duplicates of a same parameter are found, to avoid HTTP Parameter Pollution like vulnerabilities.

Required Associated Types

type Decoder: Iterator + 'a

Required Methods

fn decode(&'a self) -> Self::Decoderwhere <Self::Decoder as Iterator>::Item: Parameter,

Returns this selector’s parameters as an iterator.

Provided Methods

fn decode_into_map<N, V>( &'a self ) -> Result<HashMap<N, V, RandomState>, Box<dyn Error + Send + Sync + 'static, Global>>where <Self::Decoder as Iterator>::Item: Parameter, N: AsRef<str> + Hash + Eq, <<Self::Decoder as Iterator>::Item as Parameter>::Name: Into<N>, <<Self::Decoder as Iterator>::Item as Parameter>::Value: Into<V>,

Extracts all parameters into a HashMap, returning an error if duplicate parameters arrise.

fn get_parameters<const N: usize>( &'a self, names: [&str; N] ) -> Result<[Option<<<Self::Decoder as Iterator>::Item as Parameter>::Value>; N], Box<dyn Error + Send + Sync + 'static, Global>>where <Self::Decoder as Iterator>::Item: Parameter,

Extracts the requested parameters from the selector parameters.

The default implementation is done in a single pass through the selector parameters, returning an error if any of the requested parameters are present more than once.

fn get_bools<const N: usize>( &'a self, names: [&str; N] ) -> Result<[bool; N], Box<dyn Error + Send + Sync + 'static, Global>>where <Self::Decoder as Iterator>::Item: Parameter,

Extracts the requested arguments from the selector parameters as booleans, following the Zenoh convention that if a parameter name is present and has a value different from “false”, its value is truthy.

The default implementation is done in a single pass through the selector parameters, returning an error if some of the requested parameters are present more than once.

fn time_range( &'a self ) -> Result<Option<TimeRange<TimeExpr>>, Box<dyn Error + Send + Sync + 'static, Global>>where <Self::Decoder as Iterator>::Item: Parameter,

Extracts the standardized _time argument from the selector parameters.

The default implementation still causes a complete pass through the selector parameters to ensure that there are no duplicates of the _time key.