Trait space_net::node::SessionDeclarations

pub trait SessionDeclarations {
    // Required methods
    fn declare_subscriber<'a, TryIntoKeyExpr>(
        &self,
        key_expr: TryIntoKeyExpr
    ) -> SubscriberBuilder<'static, 'a, PushMode, DefaultHandler>
       where TryIntoKeyExpr: TryInto<KeyExpr<'a>>,
             <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>;
    fn declare_queryable<'a, TryIntoKeyExpr>(
        &self,
        key_expr: TryIntoKeyExpr
    ) -> QueryableBuilder<'static, 'a, DefaultHandler>
       where TryIntoKeyExpr: TryInto<KeyExpr<'a>>,
             <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>;
    fn declare_publisher<'a, TryIntoKeyExpr>(
        &self,
        key_expr: TryIntoKeyExpr
    ) -> PublisherBuilder<'static, 'a>
       where TryIntoKeyExpr: TryInto<KeyExpr<'a>>,
             <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>;
}

Functions to create zenoh entities with 'static lifetime.

This trait contains functions to create zenoh entities like Subscriber, and Queryable with a 'static lifetime. This is useful to move zenoh entities to several threads and tasks.

This trait is implemented for Arc<Session>.

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap().into_arc();
let subscriber = session.declare_subscriber("key/expression")
    .res()
    .await
    .unwrap();
async_std::task::spawn(async move {
    while let Ok(sample) = subscriber.recv_async().await {
        println!("Received : {:?}", sample);
    }
}).await;

Required Methods

fn declare_subscriber<'a, TryIntoKeyExpr>( &self, key_expr: TryIntoKeyExpr ) -> SubscriberBuilder<'static, 'a, PushMode, DefaultHandler>where TryIntoKeyExpr: TryInto<KeyExpr<'a>>, <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Subscriber for the given key expression.

Arguments

key_expr - The resourkey expression to subscribe to

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap().into_arc();
let subscriber = session.declare_subscriber("key/expression")
    .res()
    .await
    .unwrap();
async_std::task::spawn(async move {
    while let Ok(sample) = subscriber.recv_async().await {
        println!("Received : {:?}", sample);
    }
}).await;

fn declare_queryable<'a, TryIntoKeyExpr>( &self, key_expr: TryIntoKeyExpr ) -> QueryableBuilder<'static, 'a, DefaultHandler>where TryIntoKeyExpr: TryInto<KeyExpr<'a>>, <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Queryable for the given key expression.

Arguments

key_expr - The key expression matching the queries the Queryable will reply to

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap().into_arc();
let queryable = session.declare_queryable("key/expression")
    .res()
    .await
    .unwrap();
async_std::task::spawn(async move {
    while let Ok(query) = queryable.recv_async().await {
        query.reply(Ok(Sample::try_from(
            "key/expression",
            "value",
        ).unwrap())).res().await.unwrap();
    }
}).await;

fn declare_publisher<'a, TryIntoKeyExpr>( &self, key_expr: TryIntoKeyExpr ) -> PublisherBuilder<'static, 'a>where TryIntoKeyExpr: TryInto<KeyExpr<'a>>, <TryIntoKeyExpr as TryInto<KeyExpr<'a>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Publisher for the given key expression.

Arguments

key_expr - The key expression matching resources to write

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap().into_arc();
let publisher = session.declare_publisher("key/expression")
    .res()
    .await
    .unwrap();
publisher.put("value").res().await.unwrap();