Struct space_net::boot_node::Session

pub struct Session {
    pub(crate) runtime: Runtime,
    pub(crate) state: Arc<RwLock<SessionState>>,
    pub(crate) id: u16,
    pub(crate) alive: bool,
}

A zenoh session.

Fields
runtime: Runtimestate: Arc<RwLock<SessionState>>id: u16alive: bool
Implementations

impl Session

pub fn into_arc(self) -> Arc<Session>

Consumes the given Session, returning a thread-safe reference-counting pointer to it (Arc<Session>). This is equivalent to Arc::new(session).

This is useful to share ownership of the Session between several threads and tasks. It also alows to create Subscriber and Queryable with static lifetime that can be moved to several threads and tasks

Note: the given zenoh Session will be closed when the last reference to it is dropped.

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap().into_arc();
let subscriber = session.declare_subscriber("key/expression")
    .res()
    .await
    .unwrap();
async_std::task::spawn(async move {
    while let Ok(sample) = subscriber.recv_async().await {
        println!("Received : {:?}", sample);
    }
}).await;

pub fn leak(s: Session) -> &'static mut Session

Consumes and leaks the given Session, returning a 'static mutable reference to it. The given Session will live for the remainder of the program’s life. Dropping the returned reference will cause a memory leak.

This is useful to move entities (like Subscriber) which lifetimes are bound to the session lifetime in several threads or tasks.

Note: the given zenoh Session cannot be closed any more. At process termination the zenoh session will terminate abruptly. If possible prefer using Session::into_arc().

Examples

use zenoh::prelude::r#async::*;
use zenoh::Session;

let session = Session::leak(zenoh::open(config::peer()).res().await.unwrap());
let subscriber = session.declare_subscriber("key/expression").res().await.unwrap();
async_std::task::spawn(async move {
    while let Ok(sample) = subscriber.recv_async().await {
        println!("Received : {:?}", sample);
    }
}).await;

pub fn zid(&self) -> ZenohId

Returns the identifier of the current session. zid() is a convenient shortcut. See Session::info() and SessionInfo::zid() for more details.

pub fn hlc(&self) -> Option<&HLC>

pub fn close( self ) -> impl Resolve<Result<(), Box<dyn Error + Send + Sync + 'static, Global>>>

Close the zenoh Session.

Sessions are automatically closed when dropped, but you may want to use this function to handle errors or close the Session asynchronously.

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
session.close().res().await.unwrap();

pub fn undeclare<'a, T, O>(&'a self, decl: T) -> Owhere O: Resolve<Result<(), Box<dyn Error + Send + Sync + 'static, Global>>>, T: Undeclarable<&'a Session, O, Result<(), Box<dyn Error + Send + Sync + 'static, Global>>>,

pub fn config(&self) -> &Notifier<Config>

Get the current configuration of the zenoh Session.

The returned configuration Notifier can be used to read the current zenoh configuration through the get function or modify the zenoh configuration through the insert, or insert_json5 funtion.

Examples

Read current zenoh configuration

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let peers = session.config().get("connect/endpoints").unwrap();

Modify current zenoh configuration

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let _ = session.config().insert_json5("connect/endpoints", r#"["tcp/127.0.0.1/7447"]"#);

pub fn info(&self) -> SessionInfo<'_>

Get informations about the zenoh Session.

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let info = session.info();

pub fn declare_subscriber<'b, TryIntoKeyExpr, 'a>( &'a self, key_expr: TryIntoKeyExpr ) -> SubscriberBuilder<'a, 'b, PushMode, DefaultHandler>where TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Subscriber for the given key expression.

Arguments

key_expr - The key expression to subscribe to

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let subscriber = session.declare_subscriber("key/expression").res().await.unwrap();
while let Ok(sample) = subscriber.recv_async().await {
    println!("Received : {:?}", sample);
}

pub fn declare_queryable<'b, TryIntoKeyExpr, 'a>( &'a self, key_expr: TryIntoKeyExpr ) -> QueryableBuilder<'a, 'b, DefaultHandler>where TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Queryable for the given key expression.

Arguments

key_expr - The key expression matching the queries the Queryable will reply to

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let queryable = session.declare_queryable("key/expression").res().await.unwrap();
while let Ok(query) = queryable.recv_async().await {
    query.reply(Ok(Sample::try_from(
        "key/expression",
        "value",
    ).unwrap())).res().await.unwrap();
}

pub fn declare_publisher<'b, TryIntoKeyExpr, 'a>( &'a self, key_expr: TryIntoKeyExpr ) -> PublisherBuilder<'a, 'b>where TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Create a Publisher for the given key expression.

Arguments

key_expr - The key expression matching resources to write

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let publisher = session.declare_publisher("key/expression")
    .res()
    .await
    .unwrap();
publisher.put("value").res().await.unwrap();

pub fn declare_keyexpr<'a, 'b, TryIntoKeyExpr>( &'a self, key_expr: TryIntoKeyExpr ) -> impl Resolve<Result<KeyExpr<'b>, Box<dyn Error + Send + Sync + 'static, Global>>> + 'awhere 'b: 'a, TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Informs Zenoh that you intend to use key_expr multiple times and that it should optimize its transmission.

The returned KeyExpr’s internal structure may differ from what you would have obtained through a simple key_expr.try_into(), to save time on detecting the optimizations that have been associated with it.

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let key_expr = session.declare_keyexpr("key/expression").res().await.unwrap();

pub fn put<'a, 'b, TryIntoKeyExpr, IntoValue>( &'a self, key_expr: TryIntoKeyExpr, value: IntoValue ) -> PutBuilder<'a, 'b>where 'b: 'a, TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>, IntoValue: Into<Value>,

Put data.

Arguments

key_expr - Key expression matching the resources to put

value - The value to put

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
session
    .put("key/expression", "value")
    .encoding(KnownEncoding::TextPlain)
    .res()
    .await
    .unwrap();

pub fn delete<'a, 'b, TryIntoKeyExpr>( &'a self, key_expr: TryIntoKeyExpr ) -> PutBuilder<'a, 'b>where 'b: 'a, TryIntoKeyExpr: TryInto<KeyExpr<'b>>, <TryIntoKeyExpr as TryInto<KeyExpr<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Delete data.

Arguments

key_expr - Key expression matching the resources to delete

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
session.delete("key/expression").res().await.unwrap();

pub fn get<'a, 'b, IntoSelector>( &'a self, selector: IntoSelector ) -> GetBuilder<'a, 'b, DefaultHandler>where 'b: 'a, IntoSelector: TryInto<Selector<'b>>, <IntoSelector as TryInto<Selector<'b>>>::Error: Into<Box<dyn Error + Send + Sync + 'static, Global>>,

Query data from the matching queryables in the system.

Unless explicitly requested via [GetBuilder::accept_replies], replies are guaranteed to have key expressions that match the requested selector.

Arguments

selector - The selection of resources to query

Examples

use zenoh::prelude::r#async::*;

let session = zenoh::open(config::peer()).res().await.unwrap();
let replies = session.get("key/expression").res().await.unwrap();
while let Ok(reply) = replies.recv_async().await {
    println!(">> Received {:?}", reply.sample);
}